+++ grails-app/services/io/xh/toolbox/user/RoleService.groovy
@@ -13,23 +13,8 @@ import static io.xh.hoist.util.InstanceConfigUtils.getInstanceConfig
  */
 class RoleService extends DefaultRoleService {
 
-    void init() {
-        super.init()
-        assignLocalAdminRolesIfNeeded()
-    }
-
     /** Toolbox does not use any external directory (such as LDAP/AD) for group membership. */
     boolean getDirectoryGroupsSupported() {
         return false
-    }
-
-    @Transactional
-    private void assignLocalAdminRolesIfNeeded() {
-        String adminUsername = getInstanceConfig('adminUsername')
-        if (adminUsername) {
-            def user = User.findByEmail(adminUsername)
-            assignRole(user, 'HOIST_ADMIN')
-            assignRole(user, 'HOIST_ROLE_MANAGER')
-        }
     }
 }