+++ CHANGELOG.md
@@ -3,8 +3,14 @@
 ## v4.0-SNAPSHOT - unreleased
 
 ### New Features
-
+* Adds new `bootstrapLogin` app that uses `bootstrapAdminUser` and `bootstrapAdminPassword`to bypass
+`Auth0` login in development mode when `Auth0` itself isn't working or when running the app via 
+`yarn startOnDevice`.
 * Added `package.json` script `buildAndRun` for easier testing of prod build.
+
+### Technical
+* Remove obsolete `useOAuth` and `adminUsername` instanceConfigs.  Simplify bootstrapping 
+process to always use `Auth0` with `bootstrapAdminUser` for initial login.
 
 ### Libraries